tasks

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -5,7 +5,7 @@
 */
 int main(void)
 {
-	unsigned long long a = 1, b = 2, i, c, sum = 0;
+	unsigned long int a = 1, b = 2, i, c, sum = 0;
 
 	for (i = 0; i <= 4000000; i++)
 	{
